requirements.txt
alembic==1.15.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
azure-common==1.1.28 ; python_version >= "3.11" and python_version < "4.0"
azure-communication-email==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
azure-core==1.34.0 ; python_version >= "3.11" and python_version < "4.0"
azure-functions==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
azure-identity==1.22.0 ; python_version >= "3.11" and python_version < "4.0"
azure-mgmt-core==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
azure-storage-blob==12.25.1 ; python_version >= "3.11" and python_version < "4.0"
azure-storage-queue==12.12.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.4.26 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "4.0"
cryptography==44.0.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.2 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
msal==1.32.3 ; python_version >= "3.11" and python_version < "4.0"
msrest==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.2.5 ; python_version >= "3.11" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.4 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.40 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
wrlc-alma-api-client==0.1.3 ; python_version >= "3.11" and python_version < "4.0"
xmltodict==0.14.2 ; python_version >= "3.11" and python_version < "4.0"
